requirements.txt
google_auth_oauthlib==0.4.2
openpyxl==3.0.5
pandas==1.1.3
requests==2.5.0
google_api_python_client==1.12.8
protobuf==3.14.0
